refactor imageMethods to use new fetch methods
- remove unsplash method
